embedder/requirements.txt
pdfplumber>=0.10
python-docx>=1.1
python-pptx>=0.6.23
openpyxl>=3.1
xlrd>=2.0
requests>=2.31
chromadb>=0.5
